Run service as particular user for systemd and launchd (#11)
It's important to be able to specify which user the service runs as, particularly in a Linux-based
environment. Therefore, the `ServiceInstallCtx` is extended with an optional `username` field.

It's also possible to run services on Windows with a non-Administrator user account. Initial support is included for macOS and Linux.

New system tests were provided, with the base test being extended to check if the service process is
running as the user specified in the `ServiceInstallCtx` definition. It just does this by checking
the service definition.

Running with a different user only applies to system-wide services, and we now use the
`plist` crate rather than a hard coded string.