Provide the WinSW service manager
WinSW is a useful tool that can wrap any application as a Windows service; it provides the
boilerplate code for interacting with the service infrastructure on Windows. You define a
configuration for your application service in XML, then use WinSW, along with your configuration, as
a service controller.

The full specification for the service configuration can be found here:
https://github.com/winsw/winsw/blob/v3/samples/complete.xml

In this initial implementation, we provide support for most of the available elements, with the
exception of:

* serviceaccount
* startarguments (I don't know what the difference is between that and 'arguments')
* workingdirectory
* logpath and log
* env
* download
* extensions

Support for workingdirectory and env will follow in subsequent commits; these fields can be added to
the generic `ServiceInstallCtx` since they are common options on all platforms.

Most of the code here deals with building the configuration file. This code was written as a
separate function so that the different options could be unit tested. There is an integration/system
test in a similar fashion to the sc manager.