Extend `ServiceInstallCtx` with new options
It is very common to define a working directory for a service process and a list of environment
variables to pass to the process when it runs. The `ServiceInstallCtx` is extended for this because
it applies to most service managers.

In this particular case, three of the managers have been extended to support the variables: Systemd,
Launchd and WinSW. The Windows `sc.exe` service manager does not support specifying either. The
other two unix-based service managers might, but I am not familiar with those at the present time.
It should be easy enough to extend them with a further PR.