Provide default implementation for WinSwConfig
If this is not provided, when a new WinSW service manager is created, `service_definition_dir_path`
will be set to an empty `PathBuf`, meaning the current directory will be used to store the service
definitions.

The intention was for the service definitions to be stored at C:\ProgramData\service-manager.