WinSW uninstall deletes service directories (#24)
With the WinSW service manager, each service definition has its own directory. In addition to the
service definition, WinSW also produces logs that get output to this directory. In our product, we
frequently create and delete services with the same names, so we found it problematic for these
directories to remain around.

The service manager is now updated to delete the service directory upon uninstall.