Obtain WinSW location from environment variable
If WinSW is not at any of the locations specified by the system or user `Path` variable, check the
`WINSW_PATH` variable for a potential location.

This is useful because it enables putting the WinSW binary in a location that is not on `Path`. On
Windows, by default there are no locations on `Path` the user can write to without administrative
privileges. Callers of the crate can obtain the WinSW binary and put it in any location they wish,
then set `WINSW_PATH` in the same process, and not have to worry about their users not having the
binary in a `Path` location.