Introduce `autostart` field on `ServiceInstallCtx` (#25)
* Obtain WinSW location from environment variable

If WinSW is not at any of the locations specified by the system or user `Path` variable, check the
`WINSW_PATH` variable for a potential location.

This is useful because it enables putting the WinSW binary in a location that is not on `Path`. On
Windows, by default there are no locations on `Path` the user can write to without administrative
privileges. Callers of the crate can obtain the WinSW binary and put it in any location they wish,
then set `WINSW_PATH` in the same process, and not have to worry about their users not having the
binary in a `Path` location.

* Introduce `autostart` field on `ServiceInstallCtx`

BREAKING CHANGE: users who upgrade will need to explicitly add an `autostart` field into their
`ServiceInstallCtx` definition.

This controls whether a service should automatically start upon rebooting the OS. It's an option
common to all service managers and it's useful for developers to think about whether their services
should automatically start up, which I think justifies the breaking change. If the service is
resource intensive or uses a lot of bandwidth, some users actually don't want automatic start
because it can potentially render their machine unusable.

It could be that rc.d needs a little bit of additional work here. I am not very familiar with this
system.