feat!: provide generic `RestartPolicy` enum
I found myself in a situation where I need to use the 'always' and 'on success' restart policies on
Linux Systemd and macOS Launchd. It was not possible to do this while using the generic
`ServiceManager` trait; I would have to use specific service managers, which would necessitate the
introduction of conditional compilation.

Now we introduce a generic `RestartPolicy` enum to work with the generic `ServiceManager` trait.
This should cover the most important cases supported by most service managers: 'never', 'always',
and 'on failure', plus the option for 'on success'. Users can still instantiate a particular service
manager if they want to use specific restart options that are only supported by that manager.

In some cases these variants don't map directly to some service manager options, so an approximation
is used and a warning is emitted; however, it should still be very acceptable to configure the
service in this way since the behaviour should not be too surprising.

See the changelog entry for more details.

BREAKING CHANGE: the `disable_restart_on_failure` field was removed in favour of the restart policy
and the fields on specific service managers were made optional.