fix: use correct logic for launchd restart policies
For the `OnSuccess` case, the logic was actually inverted, so the service was restarting when the
exit was *not* successful. This has now been corrected.

The logic we were previously using for the `OnSuccess` case has now been applied to the `OnFailure`
case, so we now support that scenario rather than emitting a warning saying we don't support it.

A comment clarifies that when the `SuccessfulExit` construct is *not* used, this is effectively the
`Always` policy.