feat: prevent launchd services starting when keepalive is used
When the service definition includes the `KeepAlive` setting it causes services to automatically
start when `launchctl load` is used. This is quite unintuitive and is not how services behave on
other platforms.

We now make use of the `Disabled` setting to prevent the service starting automatically and make it
behave like service managers on other platforms. This is desirable when you want to add many
services first and not have them all start at the same time, which is the case for our application.
It also seems to be the behaviour most of our users expect and I personally think makes most sense.