Provide More Intuitive Behaviour for Launchd Services  (#40)
* fix: use correct logic for launchd restart policies

For the `OnSuccess` case, the logic was actually inverted, so the service was restarting when the
exit was *not* successful. This has now been corrected.

The logic we were previously using for the `OnSuccess` case has now been applied to the `OnFailure`
case, so we now support that scenario rather than emitting a warning saying we don't support it.

A comment clarifies that when the `SuccessfulExit` construct is *not* used, this is effectively the
`Always` policy.

* feat: prevent launchd services starting when keepalive is used

When the service definition includes the `KeepAlive` setting it causes services to automatically
start when `launchctl load` is used. This is quite unintuitive and is not how services behave on
other platforms.

We now make use of the `Disabled` setting to prevent the service starting automatically and make it
behave like service managers on other platforms. This is desirable when you want to add many
services first and not have them all start at the same time, which is the case for our application.
It also seems to be the behaviour most of our users expect and I personally think makes most sense.

* chore(release): bump to version 0.10.0